Change "invoke" to "history" for all files
