--- src/main/java/seedu/address/logic/commands/exceptions/InvokeCommandException.java
@@ -1,11 +1,0 @@
-package seedu.address.logic.commands.exceptions;
-
-/**
- * Represents the execution of a {@link InvokeCommand}.
- */
-public class InvokeCommandException extends CommandException {
-    public InvokeCommandException(String message) {
-        super(message);
-    }
-
-}
