+++ src/main/java/seedu/address/ui/CommandBox.java
@@ -6,7 +6,7 @@ import javafx.scene.control.TextField;
 import javafx.scene.layout.Region;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.commands.exceptions.InvokeCommandException;
+import seedu.address.logic.commands.exceptions.HistoryCommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
@@ -45,7 +45,7 @@ public class CommandBox extends UiPart<Region> {
         try {
             commandExecutor.execute(commandText);
             commandTextField.setText("");
-        } catch (InvokeCommandException e) {
+        } catch (HistoryCommandException e) {
             commandTextField.setText(e.getMessage());
         } catch (CommandException | ParseException e) {
             setStyleToIndicateCommandFailure();
